events can be read from db
